TECH: Restructure and clean
