+++ src/react/features/filters/text/useTextFilter.js
