+++ src/react/features/navigation/actions.js
@@ -5,7 +5,6 @@ import {
   SET_FOCUS,
   GO_HOME,
   GO_SIDE,
-  SET_FULLSCREEN,
   SET_SELECTED_LINK,
 } from './actionsTypes'
 
@@ -17,11 +16,6 @@ export const enterRoom = ({ id }) => ({
 export const setFocus = ({ focus }) => ({
   type: SET_FOCUS,
   payload: { focus },
-})
-
-export const setFullscreen = (fullscreen) => ({
-  type: SET_FULLSCREEN,
-  payload: { fullscreen },
 })
 
 export const goBackward = () => ({ type: GO_BACKWARD })
