+++ src/react/features/navigation/actionsTypes.js
@@ -1,6 +1,5 @@
 export const ENTER_ROOM = 'ENTER_ROOM'
 export const SET_FOCUS = 'SET_FOCUS'
-export const SET_FULLSCREEN = 'SET_FULLSCREEN'
 export const GO_BACKWARD = 'GO_BACKWARD'
 export const GO_FORWARD = 'GO_FORWARD'
 export const GO_HOME = 'GO_HOME'
